alsa-lib:pcm: allow users to configure different period sizes.
This patch allows the effective period size to be a multiple of the slave-pcm period size.
Allowing only exact multiple of original period size is achieved by borrowing code from the
kernel hwrules implemention.

This patch is intended to save cpu workload when for example, the slave operates with very
small periods but a user does not need that small periods.

This feature is enabled by default and can be disabled by adding config option 'var_periodsize 0'.

Signed-off-by: Alexander Jahn <ajahn@de.adit-jv.com>
Signed-off-by: Andreas Pape <apape@de.adit-jv.com>