alsa-lib: pcm:dshare enable silence
This issue depends on system load - if the process using dshare is scheduled
fast enough, then there is no noise. A delay of e.g >~2ms produces hearable noise.

Reproduction with instrumented aplay(sleep every 100th period for a given time):
During the sleep time of 2000000us (2s) the hardware plays old samples in a
loop before xrun is detected and recovered after the sleep.
This is resolved by placing it in silence, in case of dshare plugin.

Signed-off-by: Alexander Jahn <ajahn@de.adit-jv.com>
Signed-off-by: Andreas Pape <apape@de.adit-jv.com>